moved take screenshot to main.js to prevent conflicts with VR application
